Move to Network
Move to network namespace. Get rid of q() and try to get used to [ ] instead of array()
